+++ src/main/java/com/liberty/repository/BookCollectionRepository.java
@@ -19,5 +19,5 @@ public interface BookCollectionRepository extends JpaRepository<BookCollectionEn
     List<BookCollectionEntity> findAllByBookId(Long bookId);
 
     @Query(nativeQuery = true, value = "SELECT * FROM libseq WHERE BookId IN (SELECT BookId FROM libavtor WHERE AvtorId= :authorId)")
-    List<BookCollectionEntity> findAllByAuthor(@Param("authorId") Integer authorId);
+    List<BookCollectionEntity> findAllByAuthor(@Param("authorId") Long authorId);
 }
